Verify fixes
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@425 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/DownloadWithTiger.cpp
@@ -542,7 +542,7 @@ void CDownloadWithTiger::FinishValidation()
 			m_pTigerBlock[ m_nVerifyBlock ] = TS_FALSE;
 			
 			pCorrupted = CFileFragment::New();
-			pCorrupted->m_nOffset	= m_nVerifyBlock * m_nTigerSize;
+			pCorrupted->m_nOffset	= (QWORD)m_nVerifyBlock * (QWORD)m_nTigerSize;
 			pCorrupted->m_nLength	= min( m_nTigerSize, m_nSize - pCorrupted->m_nOffset );
 		}
 	}
@@ -558,7 +558,7 @@ void CDownloadWithTiger::FinishValidation()
 			m_pHashsetBlock[ m_nVerifyBlock ] = TS_FALSE;
 			
 			pCorrupted = CFileFragment::New();
-			pCorrupted->m_nOffset	= m_nVerifyBlock * ED2K_PART_SIZE;
+			pCorrupted->m_nOffset	= (QWORD)m_nVerifyBlock * ED2K_PART_SIZE;
 			pCorrupted->m_nLength	= min( ED2K_PART_SIZE, m_nSize - pCorrupted->m_nOffset );
 		}
 	}
@@ -576,7 +576,7 @@ void CDownloadWithTiger::FinishValidation()
 			m_pTorrentBlock[ m_nVerifyBlock ] = TS_FALSE;
 			
 			pCorrupted = CFileFragment::New();
-			pCorrupted->m_nOffset	= m_nVerifyBlock * m_nTorrentSize;
+			pCorrupted->m_nOffset	= (QWORD)m_nVerifyBlock * (QWORD)m_nTorrentSize;
 			pCorrupted->m_nLength	= min( m_nTorrentSize, m_nSize - pCorrupted->m_nOffset );
 		}
 	}
